refactor: ajusta namespace de objetos Result
