+++ JwtStore.Core.UnitTests/SharedContext/Primitives/Results/ResultTests.cs
@@ -1,8 +1,8 @@
 using FluentAssertions;
 using JwtStore.Core.SharedContext.Primitives;
-using JwtStore.Core.SharedContext.Primitives.Result;
+using JwtStore.Core.SharedContext.Primitives.Results;
 
-namespace JwtStore.Core.UnitTests.SharedContext.Primitives.Result;
+namespace JwtStore.Core.UnitTests.SharedContext.Primitives.Results;
 
 public class ResultTests
 {
@@ -10,7 +10,7 @@ public class ResultTests
     public void InitializesFailureResult()
     {
         var error = new Error("error", "error message");
-        var result = Core.SharedContext.Primitives.Result.Result.Failure(error);
+        var result = Result.Failure(error);
 
         result.IsSuccess.Should().BeFalse();
         result.Status.Should().Be(ResultStatus.Error);
@@ -25,7 +25,7 @@ public class ResultTests
             new("error2", "error message")
         };
 
-        var result = Core.SharedContext.Primitives.Result.Result.Failure(errors);
+        var result = Result.Failure(errors);
 
         result.IsSuccess.Should()
             .BeFalse();
@@ -45,7 +45,7 @@ public class ResultTests
             new("error2", "error message")
         };
 
-        var result = Core.SharedContext.Primitives.Result.Result.Failure(errors);
+        var result = Result.Failure(errors);
 
         result.IsSuccess.Should().BeFalse();
         result.Status.Should().Be(ResultStatus.Error);
@@ -59,10 +59,24 @@ public class ResultTests
     [Fact]
     public void InitializesSuccessResult()
     {
-        var result = Core.SharedContext.Primitives.Result.Result.Success();
+        var result = Result.Success();
 
         result.IsSuccess.Should()
             .BeTrue();
         result.Status.Should().Be(ResultStatus.Ok);
+    }
+
+    [Fact]
+    public void InitializesSuccessResultOfValue()
+    {
+        var result = Result.Success<object>(new());
+
+        result.IsSuccess
+            .Should()
+            .BeTrue();
+
+        result.Status
+            .Should()
+            .Be(ResultStatus.Ok);
     }
 }
